modified:   mace/ops/activation.h 	modified:   mace/ops/cast.cc 	modified:   mace/ops/common/activation_type.h 	modified:   mace/ops/gather.cc 	modified:   mace/ops/matmul.cc 	modified:   mace/ops/ref/activation.cc 	modified:   mace/ops/softmax.cc 	modified:   mace/ops/unsqueeze.cc 	modified:   tools/python/convert.py 	modified:   tools/python/transform/base_converter.py 	modified:   tools/python/transform/onnx_converter.py 	modified:   tools/python/transform/transformer.py 	modified:   tools/python/utils/config_parser.py

+++ tools/python/utils/config_parser.py
@@ -385,7 +385,7 @@ def set_default_config_value(nor_subgraph, model):
         elif ModelKeys.quantize in model and model[ModelKeys.quantize] == 1:
             nor_subgraph[ModelKeys.data_type] = mace_pb2.DT_FLOAT
         else:
-            nor_subgraph[ModelKeys.data_type] = mace_pb2.DT_HALF
+            nor_subgraph[ModelKeys.data_type] = mace_pb2.DT_FLOAT
 
 
 def normalize_model_config(conf, model_output=None, org_model_dir=None):
